:recycle: migrate API layer to adapter pattern and consolidate service exports
