+++ src/entities/comment/api/index.ts
@@ -1,2 +1,2 @@
-export { CommentApiAdapter } from "./comment.adapter"
+export { commentAdapter } from "./comment.adapter"
 export { COMMENT_QUERY_KEY } from "./comment.query-key"
