+++ src/entities/comment/service/index.ts
@@ -1,0 +1,1 @@
+export { CommentMapperService } from "./mapper.service"
