+++ src/entities/user/api/index.ts
@@ -1,1 +1,2 @@
-export { userApi } from "./user.api"
+export { userAdapter } from "./user.adapter"
+export { USER_QUERY_KEY } from "./user.query-key"
