+++ src/entities/user/service/index.ts
@@ -1,0 +1,1 @@
+export { UserMapperService } from "./mapper.service"
