+++ src/features/comment/usecase/add-comment.usecase.ts
@@ -1,23 +1,23 @@
-import { CommentDto } from "@/entities/comment/dto"
 import { CommentFactory } from "@/entities/comment/core/comment.factory"
+import { CommentDto } from "@/entities/comment/dto"
 import { CommentRepository } from "@/entities/comment/repository"
-import { CommentMapperService } from "@/entities/comment/service/mapper.service"
-import { userApi } from "@/entities/user/api/user.api"
-import { apiClient } from "@/shared/api/api"
+import { CommentMapperService } from "@/entities/comment/service"
+import { UserRepository } from "@/entities/user/repository"
 
 /**
  * 새로운 댓글을 추가하는 UseCase
  */
 export class AddCommentUseCase {
-  private userApiInstance = userApi(apiClient)
-  
-  constructor(private readonly commentRepository: CommentRepository) {}
+  constructor(
+    private readonly commentRepository: CommentRepository,
+    private readonly userRepository: UserRepository,
+  ) {}
 
   async execute(body: string, postId: number, userId: number): Promise<CommentDto> {
     try {
       // 사용자 정보 가져오기
-      const user = await this.userApiInstance.getProfile(userId)
-      
+      const user = await this.userRepository.getUserProfile(userId)
+
       // 새 도메인 모델 생성
       const newComment = CommentFactory.createNew(body, postId, {
         id: userId,
